Fix Chrome v109 `event.path` deprecation
This fixes the Chrome v109 deprecation of `event.path` in the January 2023 release of Chrome.

https://chromestatus.com/feature/5726124632965120

This refactors all usage of `event.path` to a centralized method in the Helpers named `getEventTarget`. I wasn't certain if we could use the `event.target`, but I kept all functionality the same by utilizing `event.composedPath()` as pointed out in the deprecation notes.