Add gitkeep for dist/html
Cannot build w/o this directory. We should just nuke it all and create it if it doesn't exist, but this is a quick and dirty fix for now. I don't want to waste the time on that. I'm not sure if I even like this current build pipeline anyways. I might check out parcel's web extension recipe.