Remove circular dependency in NavigationRenderer
UI Looks broken on the options page but it's finally showing now.

Maybe the sass loading is not working as expected 🤔

There's also still a ton of circular dependency warnings,
but I think the compiler is handling those 🤞
it would be good to check further into those at some point though.