Create sections on the testing page
This makes it easier to see what everything is.

It seems iframes are not working; we do get the event for clicking into it, but we do not get the quick access UI. I'm also not certain forms are working as intended. I thought it's supposed to be able to restore a whole form rather than only just fields. I'm also not certain how other input types are supposed to be handled since